+++ realm/realm-library/src/main/java/io/realm/OrderedRealmCollection.java
@@ -50,7 +50,7 @@ public interface OrderedRealmCollection<E extends RealmModel> extends List<E>, R
      * @return a new sorted {@link RealmResults} will be created and returned. The original collection stays unchanged.
      * @throws java.lang.IllegalArgumentException if field name does not exist or it has an invalid type.
      * @throws java.lang.IllegalStateException if the Realm is closed, called on the wrong thread or the collection is
-     *                                         an un-managed collection.
+     *                                         an unmanaged collection.
      */
     RealmResults<E> sort(String fieldName);
 
@@ -63,7 +63,7 @@ public interface OrderedRealmCollection<E extends RealmModel> extends List<E>, R
      * @return a new sorted {@link RealmResults} will be created and returned. The original collection stays unchanged.
      * @throws java.lang.IllegalArgumentException if field name does not exist or has an invalid type.
      * @throws java.lang.IllegalStateException if the Realm is closed, called on the wrong thread or the collection is
-     *                                         an un-managed collection.
+     *                                         an unmanaged collection.
      */
     RealmResults<E> sort(String fieldName, Sort sortOrder);
 
@@ -79,7 +79,7 @@ public interface OrderedRealmCollection<E extends RealmModel> extends List<E>, R
      * @return a new sorted {@link RealmResults} will be created and returned. The original collection stays unchanged.
      * @throws java.lang.IllegalArgumentException if a field name does not exist or has an invalid type.
      * @throws java.lang.IllegalStateException if the Realm is closed, called on the wrong thread or the collection is
-     *                                         an un-managed collection.
+     *                                         an unmanaged collection.
      */
     RealmResults<E> sort(String fieldName1, Sort sortOrder1, String fieldName2, Sort sortOrder2);
 
@@ -92,7 +92,7 @@ public interface OrderedRealmCollection<E extends RealmModel> extends List<E>, R
      * @return a new sorted {@link RealmResults} will be created and returned. The original collection stays unchanged.
      * @throws java.lang.IllegalArgumentException if a field name does not exist or has an invalid type.
      * @throws java.lang.IllegalStateException if the Realm is closed, called on the wrong thread or the collection is
-     *                                         an un-managed collection.
+     *                                         an unmanaged collection.
      */
     RealmResults<E> sort(String[] fieldNames, Sort[] sortOrders);
 
@@ -102,7 +102,7 @@ public interface OrderedRealmCollection<E extends RealmModel> extends List<E>, R
      * @param location the array index identifying the object to be removed.
      * @throws IndexOutOfBoundsException if {@code location < 0 || location >= size()}.
      * @throws java.lang.IllegalStateException if the Realm is closed or the method is called from the wrong thread.
-     * @throws UnsupportedOperationException if the collection is un-managed.
+     * @throws UnsupportedOperationException if the collection is unmanaged.
      */
     void deleteFromRealm(int location);
 
@@ -111,7 +111,7 @@ public interface OrderedRealmCollection<E extends RealmModel> extends List<E>, R
      *
      * @return {@code true} if an object was deleted, {@code false} otherwise.
      * @throws java.lang.IllegalStateException if the Realm is closed or the method is called on the wrong thread.
-     * @throws UnsupportedOperationException if the collection is un-managed.
+     * @throws UnsupportedOperationException if the collection is unmanaged.
      */
     boolean deleteFirstFromRealm();
 
@@ -120,7 +120,7 @@ public interface OrderedRealmCollection<E extends RealmModel> extends List<E>, R
      *
      * @return {@code true} if an object was deleted, {@code false} otherwise.
      * @throws java.lang.IllegalStateException if the Realm is closed or the method is called from the wrong thread.
-     * @throws UnsupportedOperationException if the collection is un-managed.
+     * @throws UnsupportedOperationException if the collection is unmanaged.
      */
     boolean deleteLastFromRealm();
 }